Applying validations to DimensionController

+++ app/Http/Controllers/DimensionController.php
@@ -3,15 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Dimension;
-use Illuminate\Http\Request;
+use App\Http\Requests\CreateDimensionRequest;
+use App\Http\Requests\UpdateDimensionRequest;
 
 class DimensionController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $dimensions = Dimension::with('questions')->orderBy('title')->get();
@@ -19,53 +15,33 @@ class DimensionController extends Controller
     }
 
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function store(CreateDimensionRequest $request)
     {
         $data = $request->all();
         $dimension = Dimension::create($data);
         return response()->json($dimension);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Dimension  $dimension
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Dimension $dimension)
     {
         return response()->json($dimension);
     }
 
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Dimension  $dimension
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Dimension $dimension)
+    public function update(UpdateDimensionRequest $request, Dimension $dimension)
     {
         $data = $request->all();
         $dimension->fill($data)->save();
         return response()->json($dimension);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Dimension  $dimension
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy(Dimension $dimension)
     {
+        if ($dimension->questions->count() > 0) {
+            return response()->json(["message" => "The dimension cannot be deleted because she has questions related"], 403);
+        }
         return response()->json(['deleted' => $dimension->delete()]);
     }
 }
